requirements.txt
altair==4.1.0
appdirs==1.4.4
asn1crypto==1.2.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
base58==2.0.1
beautifulsoup4==4.9.1
bleach==3.1.0
blinker==1.4
blis==0.4.1
bokeh==2.1.1
boto3==1.14.24
botocore==1.17.24
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.1.1
cachy==0.3.0
catalogue==1.0.0
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
cleo==0.7.6
click==7.1.2
clikit==0.4.3
conda==4.7.12
conda-package-handling==1.6.0
cryptography==2.8
cycler==0.10.0
cymem==2.0.3
decorator==4.4.1
defusedxml==0.6.0
distlib==0.3.1
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
fastai2==0.0.17
fastcore==0.1.18
fastprogress==0.2.3
fastscript==0.1.4
filelock==3.0.12
Flask==1.1.2
future==0.18.2
html5lib==1.1
idna==2.8
imageio==2.9.0
importlib-metadata==1.1.3
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
jeepney==0.4.3
Jinja2==2.11.1
jmespath==0.10.0
joblib==0.16.0
json5==0.9.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.6
keyring==20.0.1
kiwisolver==1.2.0
kornia==0.3.1
lockfile==0.12.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
msgpack==1.0.0
murmurhash==1.0.2
nbconvert==5.6.1
nbdev==0.2.18
nbformat==5.0.4
networkx==2.4
nltk==3.5
notebook==6.0.3
numpy==1.19.0
opencv-python==4.3.0.36
packaging==20.4
panda==0.3.1
pandas==1.0.5
pandocfilters==1.4.2
parso==0.6.1
pastel==0.2.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkginfo==1.5.0.1
plac==1.1.3
poetry==1.0.9
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.12.2
ptyprocess==0.6.0
pyarrow==0.17.1
pycosat==0.6.3
pycparser==2.19
pydeck==0.4.0b2
pydicom==2.0.0
Pygments==2.5.2
pylev==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.14.11
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==18.1.1
regex==2020.6.8
requests==2.22.0
requests-toolbelt==0.8.0
ruamel-yaml==0.15.46
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.1
SecretStorage==3.1.2
Send2Trash==1.5.0
shellingham==1.3.2
six==1.12.0
soupsieve==2.0.1
spacy==2.3.2
srsly==1.0.2
streamlit==0.63.1
terminado==0.8.3
testpath==0.4.4
thinc==7.4.1
threadpoolctl==2.1.0
tifffile==2020.7.4
toml==0.10.1
tomlkit==0.5.11
toolz==0.10.0
torch==1.5.0
torchvision==0.6.1
tornado==6.0.3
tqdm==4.47.0
traitlets==4.3.3
typing-extensions==3.7.4.2
tzlocal==2.1
urllib3==1.24.2
validators==0.16.0
virtualenv==20.0.26
wasabi==0.7.0
watchdog==0.10.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
xlrd==1.2.0
zipp==2.2.0